Added createTicket() in main.js
Added createTicket() function in main.js, then modified createForm.html to add an onclick="createTicket()" attribute to the Submit button.  The onclick causes the createTicket() function to be called, which makes a POST request to the API.  Also added onsubmit="return(false)" attribute to the <form> tag in createForm.html.  Without that attribute, HTML will try to submit that form if the user clicks a submit button.  We don't want it to reload the page, so we just use JS to get the values from the input fields and send them in the body of a POST request to the API.